OPTM-49 Moved Packet to messages folder
